Fix the width of expander in wxDataViewCtrl
In 8c081b33794a8bf48ad5f14be247350778ddbe82, the width of a tree
expander is changed to the native value on the platform. But some
calculations still use guessed value (m_lineHeight) as the width. This
fixes the value, by retrieving the width using the same method as in
OnPaint().

Closes https://github.com/wxWidgets/wxWidgets/pull/1508

Closes #18473.